Move Swift & Objective C to /components
